Base Smell Sensory Stuff
This adds in the sensory manager and sense of smell for the rat. It changes the color of contaminables in it's radius which is affected by the ratstats. There are no contexts where it changes but there is a way to change it from the sensory manager. Contaminable must have mesh renderer on whatever game object the contaminable script is on