Hearing stuff
The rat now emits sound, this is basic right now but has the framework to be changed to how we like. Anything with the "SenseOfHearing" script will be able to hear the rat and anything that uses the game manager's emit sound as well will be able to be heard.